+++ scylla/src/statement/prepared_statement.rs
@@ -1,5 +1,4 @@
 use bytes::{Bytes, BytesMut};
-use scylla_cql::errors::BadQuery;
 use scylla_cql::frame::response::result::{
     ColumnSpec, PartitionKeyIndex, ResultMetadata, TableSpec,
 };
@@ -14,7 +13,7 @@ use thiserror::Error;
 use uuid::Uuid;
 
 use super::{PageSize, StatementConfig};
-use crate::errors::QueryError;
+use crate::errors::{BadQuery, QueryError};
 use crate::frame::response::result::PreparedMetadata;
 use crate::frame::types::{Consistency, SerialConsistency};
 use crate::history::HistoryListener;
